Refactor is_fpga_kernel outside of codegen
